+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 - Add `find-doc` REPL function ([#141](https://github.com/anmonteiro/lumo/issues/141)).
+- Add `source` REPL special ([#84](https://github.com/anmonteiro/lumo/issues/84)).
 
 ## [1.4.1](https://github.com/anmonteiro/lumo/compare/1.4.0...1.4.1) (2017-04-20)